More robust bleaching and merge
Bleaching regularly is of critical importance for network functionality
and correct convergence, so move periodical bleach into its own loop,
with more safeguards to deal with slowdown and high load.
This should prevent old data appearing as fresh and hence polluting the
network due to other operations taking longer then expected.
When merging data from a remote node, entries authored by this node
should always be fresher on the current node, while the opposite can be
true when merging from a CLI instance on the same machine that is
inserting data. Merge is now more robust also when the network degradate
temporarly and as a side effect insert is not rate limited anymore.